pyproject.toml
[tool.poetry]
name = "verta-chat"
version = "0.1.0"
description = ""
authors = ["Henil Gajjar <henilgajjar08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
beautifulsoup4 = "^4.12.3"
faiss-cpu = "^1.9.0"
langchain-community = "^0.3.7"
langchain-openai = "^0.2.8"
wandb = "^0.18.7"
pypdf = "^5.1.0"
jq = "^1.8.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.8.0"
langfuse = "^2.53.9"
streamlit = "^1.40.1"
ragas = "^0.1.14"
nest-asyncio = "^1.6.0"
scikit-learn = "^1.5.2"
ipykernel = "^6.29.5"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
langgraph = "^0.2.48"
google-cloud-bigquery = "^3.27.0"
google-api-python-client = "^2.153.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
db-dtypes = "^1.3.1"
cloud-sql-python-connector = "^1.13.0"
pg8000 = "^1.31.2"
sqlalchemy = "^1.4"
mlflow = "^2.17.2"
nltk = "^3.9.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"